Refactor ComponentFormat
ComponentFormat parsing and rust type resolving logic was scattered two
places. This PR will refactor this logic under same type.
