+++ utoipa-gen/src/component_type.rs
@@ -1,5 +1,7 @@
+use proc_macro2::TokenStream;
 use proc_macro_error::abort_call_site;
 use quote::{quote, ToTokens};
+use syn::{parse::Parse, Error, Ident, TypePath};
 
 /// Tokenizes OpenAPI data type correctly according to the Rust type
 pub struct ComponentType<'a>(pub &'a syn::TypePath);
@@ -126,10 +128,50 @@ impl ToTokens for ComponentType<'_> {
     }
 }
 
-/// Tokenizes OpenAPI data type format correctly by given Rust type.
-pub struct ComponentFormat<'a>(pub(crate) &'a syn::TypePath);
+/// Either Rust type component variant or enum variant component variant.
+#[cfg_attr(feature = "debug", derive(Debug))]
+pub enum ComponentFormat<'c> {
+    /// [`utoipa::openapi::shcema::ComponentFormat`] enum variant component format.
+    Variant(Variant),
+    /// Rust type component format.
+    Type(Type<'c>),
+}
 
 impl ComponentFormat<'_> {
+    pub fn is_known_format(&self) -> bool {
+        match self {
+            Self::Type(ty) => ty.is_known_format(),
+            Self::Variant(_) => true,
+        }
+    }
+}
+
+impl<'a> From<&'a TypePath> for ComponentFormat<'a> {
+    fn from(type_path: &'a TypePath) -> Self {
+        Self::Type(Type(type_path))
+    }
+}
+
+impl Parse for ComponentFormat<'_> {
+    fn parse(input: syn::parse::ParseStream) -> syn::Result<Self> {
+        Ok(Self::Variant(input.parse()?))
+    }
+}
+
+impl ToTokens for ComponentFormat<'_> {
+    fn to_tokens(&self, tokens: &mut TokenStream) {
+        match self {
+            Self::Type(ty) => ty.to_tokens(tokens),
+            Self::Variant(variant) => variant.to_tokens(tokens),
+        }
+    }
+}
+
+/// Tokenizes OpenAPI data type format correctly by given Rust type.
+#[cfg_attr(feature = "debug", derive(Debug))]
+pub struct Type<'a>(&'a syn::TypePath);
+
+impl Type<'_> {
     /// Check is the format know format. Known formats can be used within `quote! {...}` statements.
     pub fn is_known_format(&self) -> bool {
         let last_segment = match self.0.path.segments.last() {
@@ -170,7 +212,7 @@ fn is_known_format(name: &str) -> bool {
     )
 }
 
-impl ToTokens for ComponentFormat<'_> {
+impl ToTokens for Type<'_> {
     fn to_tokens(&self, tokens: &mut proc_macro2::TokenStream) {
         let last_segment = self.0.path.segments.last().unwrap_or_else(|| {
             abort_call_site!("expected there to be at least one segment in the path")
@@ -191,5 +233,89 @@ impl ToTokens for ComponentFormat<'_> {
             "Uuid" => tokens.extend(quote! { utoipa::openapi::ComponentFormat::Uuid }),
             _ => (),
         }
+    }
+}
+
+/// [`Parse`] and [`ToTokens`] implementation for [`utoipa::openapi::schema::ComponentFormat`].
+#[cfg_attr(feature = "debug", derive(Debug))]
+pub enum Variant {
+    Int32,
+    Int64,
+    Float,
+    Double,
+    Byte,
+    Binary,
+    Date,
+    DateTime,
+    Password,
+    #[cfg(feature = "uuid")]
+    Uuid,
+}
+
+impl Parse for Variant {
+    fn parse(input: syn::parse::ParseStream) -> syn::Result<Self> {
+        const FORMATS: [&str; 10] = [
+            "Int32", "Int64", "Float", "Double", "Byte", "Binary", "Date", "DateTime", "Password",
+            "Uuid",
+        ];
+        let allowed_formats = FORMATS
+            .into_iter()
+            .filter(|_format| {
+                #[cfg(feature = "uuid")]
+                {
+                    true
+                }
+                #[cfg(not(feature = "uuid"))]
+                {
+                    _format != &"Uuid"
+                }
+            })
+            .collect::<Vec<_>>();
+        let expected_formats = format!(
+            "unexpected format, expected one of: {}",
+            allowed_formats.join(", ")
+        );
+        let format = input.parse::<Ident>()?;
+        let name = &*format.to_string();
+
+        match name {
+            "Int32" => Ok(Self::Int32),
+            "Int64" => Ok(Self::Int64),
+            "Float" => Ok(Self::Float),
+            "Double" => Ok(Self::Double),
+            "Byte" => Ok(Self::Byte),
+            "Binary" => Ok(Self::Binary),
+            "Date" => Ok(Self::Date),
+            "DateTime" => Ok(Self::DateTime),
+            "Password" => Ok(Self::Password),
+            #[cfg(feature = "uuid")]
+            "Uuid" => Ok(Self::Uuid),
+            _ => Err(Error::new(
+                format.span(),
+                format!("unexpected format: {name}, expected one of: {expected_formats}"),
+            )),
+        }
+    }
+}
+
+impl ToTokens for Variant {
+    fn to_tokens(&self, tokens: &mut TokenStream) {
+        match self {
+            Self::Int32 => tokens.extend(quote!(utoipa::openapi::schema::ComponentFormat::Int32)),
+            Self::Int64 => tokens.extend(quote!(utoipa::openapi::schema::ComponentFormat::Int64)),
+            Self::Float => tokens.extend(quote!(utoipa::openapi::schema::ComponentFormat::Float)),
+            Self::Double => tokens.extend(quote!(utoipa::openapi::schema::ComponentFormat::Double)),
+            Self::Byte => tokens.extend(quote!(utoipa::openapi::schema::ComponentFormat::Byte)),
+            Self::Binary => tokens.extend(quote!(utoipa::openapi::schema::ComponentFormat::Binary)),
+            Self::Date => tokens.extend(quote!(utoipa::openapi::schema::ComponentFormat::Date)),
+            Self::DateTime => {
+                tokens.extend(quote!(utoipa::openapi::schema::ComponentFormat::DateTime))
+            }
+            Self::Password => {
+                tokens.extend(quote!(utoipa::openapi::schema::ComponentFormat::Password))
+            }
+            #[cfg(feature = "uuid")]
+            Self::Uuid => tokens.extend(quote!(utoipa::openapi::schema::ComponentFormat::Uuid)),
+        };
     }
 }
